requirements.txt
# Web Framework & APIs
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3.0

# API Documentation (Swagger)
drf-yasg>=1.21

# AI & Facial Recognition
deepface>=0.0.93
opencv-python-headless>=4.8
numpy>=1.24
tf-keras>=2.15

# Utilities & Security
qrcode[pil]>=7.4
Pillow>=10.0
python-dotenv>=1.0